+++ src/quantileregression.jl
@@ -3,7 +3,7 @@ module QuantileRegression
 export quantileregression
 
 using JuMP
-using GLPK
+using HiGHS
 
 import ..Basis:
     RegressionSetting, @extractRegressionSetting, designMatrix, responseVector, applyColumns
@@ -83,7 +83,8 @@ julia> result = quantileregression(X, foodexp, tau = 0.25)
 function quantileregression(X::Array{Float64,2}, y::Array{Float64,1}; tau::Float64 = 0.5)
     n, p = size(X)
 
-    m = JuMP.Model(GLPK.Optimizer)
+    m = JuMP.Model(HiGHS.Optimizer)
+    set_silent(m)
 
     JuMP.@variable(m, d[1:(2n)])
     JuMP.@variable(m, beta[1:p])